[Ada] Avoid crash when traversing units with -gnatd.WW debug switch
The debug switch -gnatd.WW enables extra info when traversing library units
with Walk_Library_Items, which is used in the CodePeer and GNATprove. This
routine was crashing when trying to print info about a unit with configuration
pragmas (typically an .adc file). Now fixed.

No test, as the crash only happens when a GNATprove backend is manually called
with -gnatd.WW switch. Frontend is not affected.

2018-07-16  Piotr Trojanek  <trojanek@adacore.com>

gcc/ada/

	* sem.adb (Walk_Library_Items): Skip units with configuration pragmas
	when printing debug info.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@262720 138bc75d-0d04-0410-961f-82ee72b054a4